+++ src/BasicDisplays/index.ts
@@ -1,5 +1,4 @@
 export * from './BooleanDisplay';
 export * from './DatetimeDisplay';
 export * from './RawDisplay';
-export * from './ModelItemDisplay';
-export * from './ModelListDisplay';
+export * from './ModelDisplay';
