+++ src/types/conveyor.ts
@@ -20,8 +20,7 @@ export interface FieldOptions {
 }
 
 export enum NonScalarType {
-  MODEL_ITEM = '__MODEL_ITEM__',
-  MODEL_LIST = '__MODEL_LIST__',
+  MODEL = '__MODEL__',
 }
 
 export const FieldType = { ...ScalarType, ...NonScalarType };
